{"schema":"decodex/commit/1","summary":"Split knowledge watch state derivation into a sibling module after strict validation.","authority":"manual"}
